+++ examples/data-sources/nxos_pim_instance/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_pim_instance" "example" {
+}