+++ examples/resources/nxos_pim_instance/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_pim_instance.example "sys/pim/inst"