+++ examples/resources/nxos_pim_instance/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_pim_instance" "example" {
+  admin_state = "enabled"
+}