+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Add nxos_ospf_interface resource and data source
 - Add nxos_ospf_authentication resource and data source
 - Add nxos_pim resource and data source
+- Add nxos_pim_instance resource and data source
 
 ## 0.2.3